Port VCL file to carlosabalde/libvmod-redis
This doesn't use the redis pipeline features, but we can approximate
it with batched GET commands anyway, so nothing's really lost.
As part of this I also tightened up dealing with failed calls to
Redis and missing values, so that throttling should still work if
you forget to set up the defaults.